feat: can add the whitelist directly (#121)

+++ web/src/views/PcHome.vue
@@ -419,6 +419,9 @@ const handleWhiteList = () => {
   }
 };
 const getWhiteList = async () => {
+  if (!checkAuthToken()) {
+    return;
+  }
   const { data } = await new ApiService().getWhitelist();
   if (data.value) {
     whiteList.value = data.value;
@@ -428,6 +431,10 @@ const getWhiteList = async () => {
 };
 
 const removeWhiteList = async (player) => {
+  if(player.player_uid || player.steam_id) {
+    message.error(t("message.removewhitefail", { err: "player_uid or steam_id is required" }));
+    return;
+  }
   const { data, statusCode } = await new ApiService().removeWhitelist(player);
   if (statusCode.value === 200) {
     message.success(t("message.removewhitesuccess"));
@@ -476,6 +483,28 @@ const handleAddWhiteList = () => {
     showAddWhiteListModal.value = true;
   }
 };
+const virtualListInst = ref();
+const handleAddNewWhiteList = () => {
+  whiteList.value.unshift({
+    name: "",
+    player_uid: "",
+    steam_id: "",
+  })
+  virtualListInst.value?.scrollTo({ index: 0 });
+};
+const isWhite = (player) => {
+  if(whiteList.value.length === 0) {
+    return false;
+  }
+  return whiteList.value.some(whitelistItem => {
+    return (
+        (whitelistItem.player_uid && whitelistItem.player_uid === player.player_uid) ||
+        (whitelistItem.steam_id && whitelistItem.steam_id === player.steam_id)
+    );
+  });
+}
+
+
 // broadcast
 const showBroadcastModal = ref(false);
 const broadcastText = ref("");
@@ -607,6 +636,7 @@ onMounted(async () => {
   checkAuthToken();
   getServerInfo();
   await getPlayerList();
+  await getWhiteList();
   loading.value = false;
   setInterval(() => {
     getPlayerList(false);
@@ -797,6 +827,9 @@ onMounted(async () => {
                       <n-tag class="ml-2" type="primary" size="small" round>
                         Lv.{{ player.level }}
                       </n-tag>
+                      <n-tag class="ml-2" type="warning" size="small" round v-if="isWhite(player)">
+                        {{ $t("status.whitelist") }}
+                      </n-tag>
                       <span class="flex-1 pl-2 font-bold line-clamp-1">{{
                         player.nickname
                       }}</span>
@@ -872,6 +905,9 @@ onMounted(async () => {
                           : $t("status.offline")
                       }}</n-tag
                     >
+                    <n-tag class="ml-2" type="warning" round v-if="isWhite(playerInfo)">
+                      {{ $t("status.whitelist") }}
+                    </n-tag>
                     <n-button
                       @click="copyText(playerInfo.player_uid)"
                       class="ml-3"
@@ -1330,8 +1366,8 @@ onMounted(async () => {
     :segmented="segmented"
   >
     <div>
-      <n-empty description="什么都没有" v-if="whiteList.length==0"> </n-empty>
-      <n-virtual-list v-else style="max-height: 240px" :item-size="42" :items="whiteList">
+      <n-empty description="empty" v-if="whiteList.length==0"> </n-empty>
+      <n-virtual-list ref="virtualListInst" v-else style="max-height: 240px" :item-size="42" :items="whiteList">
         <template #default="{ item }">
           <div :key="item.player_uid" class="flex flex-col item mlr-3 mb-3" style="height: 42px">
             <n-grid >
@@ -1344,7 +1380,7 @@ onMounted(async () => {
               </n-gi>
               <n-gi span="5">
                 <div class="flex justify-end mr-3">
-                  <n-space>
+                  <n-space v-if="item.player_uid || item.steam_id">
                     <n-button
                         strong secondary type="primary"
                         @click="() => {
@@ -1377,6 +1413,13 @@ onMounted(async () => {
       <div class="flex justify-end">
         <n-space>
           <n-button
+              type="primary"
+              @click="handleAddNewWhiteList"
+          >
+            {{ $t("button.addNew") }}
+          </n-button>
+
+          <n-button
               type="tertiary"
               @click="
             () => {
@@ -1392,7 +1435,7 @@ onMounted(async () => {
               @click="putWhiteList"
               strong secondary type="success"
           >
-            保存
+            {{ $t("button.save") }}
           </n-button>
         </n-space>
       </div>
